+++ lib/amazonka-core/src/Amazonka/Error.hs
@@ -14,6 +14,7 @@ import Amazonka.Types
 import qualified Data.Aeson as Aeson
 import qualified Data.Aeson.Types as Aeson.Types
 import qualified Data.ByteString.Lazy as LBS
+import Data.Either (fromRight)
 import qualified Network.HTTP.Client as Client
 import Network.HTTP.Types.Status (Status (..))
 
@@ -94,8 +95,10 @@ serviceError a s h c m r =
   ServiceError' a s h (fromMaybe (getErrorCode s h) c) m (r <|> getRequestId h)
 
 getRequestId :: [Header] -> Maybe RequestId
-getRequestId h =
-  either (const Nothing) Just (h .# hAMZRequestId <|> h .# hAMZNRequestId)
+getRequestId h
+  | Right hAMZ  <- h .# hAMZRequestId  = Just hAMZ
+  | Right hAMZN <- h .# hAMZNRequestId = Just hAMZN
+  | otherwise                          = Nothing
 
 getErrorCode :: Status -> [Header] -> ErrorCode
 getErrorCode s h =
@@ -135,22 +138,19 @@ parseXMLError ::
   [Header] ->
   ByteStringLazy ->
   Error
-parseXMLError a s h bs = decodeError a s h bs (decodeXML bs >>= go)
+parseXMLError a s h bs = decodeError a s h bs (go <$> decodeXML bs)
   where
     go x =
       serviceError a s h
-        <$> code x
-        <*> may' (firstElement "Message" x)
-        <*> may' (firstElement "RequestId" x <|> firstElement "RequestID" x)
+        (code x)
+        (may' (firstElement "Message" x))
+        (may' (firstElement "RequestId" x) <|> may' (firstElement "RequestID" x))
 
-    code x =
-      Just <$> (firstElement "Code" x >>= parseXML)
-        <|> return root
+    code x = fromRight root $ parseXML =<< firstElement "Code" x
 
     root = newErrorCode <$> rootElementName bs
 
-    may' (Left _) = pure Nothing
-    may' (Right x) = Just <$> parseXML x
+    may' x = either (const Nothing) Just $ parseXML =<< x
 
 parseRESTError ::
   Abbrev ->
@@ -171,7 +171,7 @@ decodeError ::
 decodeError a s h bs e
   | LBS.null bs = parseRESTError a s h bs
   | otherwise =
-    either
-      (SerializeError . SerializeError' a s (Just bs))
-      ServiceError
-      e
+      either
+        (SerializeError . SerializeError' a s (Just bs))
+        ServiceError
+        e