+++ lib/amazonka/CHANGELOG.md
@@ -75,6 +75,8 @@ Released: **?**, Compare: [2.0.0-rc1](https://github.com/brendanhay/amazonka/com
 [\#723](https://github.com/brendanhay/amazonka/pull/723)
 - `amazonka-ssm`: Various fields that are always available are no longer Maybe's.
 [\#741](https://github.com/brendanhay/amazonka/pull/741)
+- `amazonka-core`: to be compatible with mtl-2.3, avoid `Alternative (Either String)`
+[\#779](https://github.com/brendanhay/amazonka/pull/779)
 
 ## [2.0.0 RC1](https://github.com/brendanhay/amazonka/tree/2.0.0-rc1)
 Released: **28nd November, 2021**, Compare: [1.6.1](https://github.com/brendanhay/amazonka/compare/1.6.1...2.0.0-rc1)